services/nlp-service/requirements.txt
azure-monitor-opentelemetry
fastapi
uvicorn[standard]
httpx
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp-proto-http


